refactor: split password functionality into separate controller
